add code

+++ River/River/Renderer/DX12Renderer/Header/DX12IndexBuffer.h
@@ -8,13 +8,13 @@
 class DX12IndexBuffer : public IndexBuffer
 {
 public:
-	DX12IndexBuffer(ID3D12Device* device, void* indices, uint32_t count, ShaderDataType indiceDataType);
+	DX12IndexBuffer(ID3D12Device* device, void* indices, uint32 count, ShaderDataType indiceDataType);
 
-	DX12IndexBuffer(ID3D12Device* device, ID3D12GraphicsCommandList* commandList, void* indices, uint32_t count, ShaderDataType indiceDataType);
+	DX12IndexBuffer(ID3D12Device* device, ID3D12GraphicsCommandList* commandList, void* indices, uint32 count, ShaderDataType indiceDataType);
 	
 	virtual ~DX12IndexBuffer() override;
 
-	virtual void UpdateData(void* context, void* cmdList, void* indices, size_t count, uint32_t additionalCount = 0) override;
+	virtual void UpdateData(void* context, void* cmdList, void* indices, uint32 count, uint32 additionalCount = 0) override;
 
 	const D3D12_INDEX_BUFFER_VIEW& GetView() const { return m_IndexBufferView; }
